moved new binding.cpp helper functions to bindings_utils.h to make git diff cleaner and easier to read
